app/requirements.txt
streamlit
elasticsearch
requests
openai
psycopg2-binary
python-dotenv
numpy
spacy